# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Unified Operations Platform API
  description: |
    Comprehensive API for the Unified Operations Platform supporting event management,
    resource planning, financial operations, and multi-tenant workflows.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /users/{userId}:
    get:
      summary: Get user
      description: Retrieve a specific user by ID
      operationId: getUser
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      summary: Update user
      description: Update an existing user
      operationId: updateUser
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete user
      description: Delete a user (soft delete)
      operationId: deleteUser
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /venues:
    get:
      summary: List venues
      description: Retrieve a paginated list of venues
      operationId: listVenues
      tags:
        - Venues
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: type
          in: query
          schema:
            type: string
          description: Filter by venue type
        - name: capacity
          in: query
          schema:
            type: integer
          description: Minimum capacity
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueListResponse"

    post:
      summary: Create venue
      description: Create a new venue
      operationId: createVenue
      tags:
        - Venues
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueRequest"
      responses:
        "201":
          description: Venue created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /venues/{venueId}:
    get:
      summary: Get venue
      description: Retrieve a specific venue by ID
      operationId: getVenue
      tags:
        - Venues
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/VenueIdParam"
      responses:
        "200":
          description: Venue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"

    patch:
      summary: Update venue
      description: Update an existing venue
      operationId: updateVenue
      tags:
        - Venues
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/VenueIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueRequest"
      responses:
        "200":
          description: Venue updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
...
  /workflows:
    get:
      summary: List workflows
      description: Retrieve a paginated list of workflow definitions
      operationId: listWorkflows
      tags:
        - Workflows
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, draft]
          description: Filter by workflow status
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowListResponse"

    post:
      summary: Create workflow
      description: Create a new workflow definition
      operationId: createWorkflow
      tags:
        - Workflows
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowRequest"
      responses:
        "201":
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Search query string
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
...
    VenueIdParam:
      name: venueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Venue ID
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
          format: uri
        role:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        organizationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
...
    Venue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        capacity:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
            format: uri
        status:
          type: string
          enum: [active, inactive, maintenance]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateVenueRequest:
      type: object
      required:
        - name
        - address
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        capacity:
          type: integer
        amenities:
          type: array
          items:
            type: string
    UpdateVenueRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        capacity:
          type: integer
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, maintenance]
    VenueListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        trigger:
          type: object
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        status:
          type: string
          enum: [active, inactive, draft]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowStep:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        config:
          type: object
        nextSteps:
          type: array
          items:
            type: string
    CreateWorkflowRequest:
      type: object
      required:
        - name
        - trigger
        - steps
      properties:
        name:
          type: string
        description:
          type: string
        trigger:
          type: object
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
    WorkflowListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Workflow"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
security:
  - bearerAuth: []